# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WebBridge Rest API discovery website.",
        "title": "WebBridge Restful API Documentation",
        "termsOfService": "http://www.duality.solutions/webbridge/terms",
        "contact": {
            "name": "API Support",
            "url": "http://www.duality.solutions/support",
            "email": "support@duality.solutions"
        },
        "license": {
            "name": "Duality",
            "url": "https://github.com/duality-solutions/web-bridge/blob/master/LICENSE.md"
        },
        "version": "1.0"
    },
    "host": "http://docs.webbridge.duality.solutions",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/ice": {
            "get": {
                "description": "Returns all the ICE servers in current running configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.IceServerConfig"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Add or update an ICE server in current configuration and saves the changes to file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "ICE Configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.IceServerConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.IceServerConfig"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Replaces an ICE server in current configuration and saves the changes to file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "ICE Configuration File",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ConfigurationFile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.ConfigurationFile"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an ICE server in current configuration and saves the changes to file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "ICE Configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.IceServerConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.IceServerConfig"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ConfigurationFile": {
            "type": "object",
            "properties": {
                "IceServers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.IceServerConfig"
                    }
                }
            }
        },
        "models.IceServerConfig": {
            "type": "object",
            "properties": {
                "Credential": {
                    "description": "The ICE server's credentials. Leave blank if it doesn't apply",
                    "type": "string"
                },
                "URL": {
                    "description": "The ICE server's full URL with protocol prefix and port: turn:ice.bdap.io:3478",
                    "type": "string"
                },
                "UserName": {
                    "description": "The ICE server's user name. Leave blank if it doesn't apply",
                    "type": "string"
                }
            }
        }
    }
}